chore: rename module internals (routing::routing -> routing::routing_api eg)
