+++ src/routing/routing_api/command.rs
@@ -11,7 +11,7 @@ use crate::messaging::{
     section_info::SectionInfoMsg,
     DstInfo, Itinerary, MessageType, SectionAuthorityProvider,
 };
-use crate::routing::{node::Node, routing::Peer, section::SectionKeyShare, XorName};
+use crate::routing::{node::Node, routing_api::Peer, section::SectionKeyShare, XorName};
 use bytes::Bytes;
 use hex_fmt::HexFmt;
 use std::{
