+++ src/routing/routing_api/core/messaging/handling/bad_msgs.rs
@@ -14,7 +14,7 @@ use crate::messaging::{
 use crate::routing::{
     messages::{RoutingMsgUtils, SrcAuthorityUtils},
     peer::PeerUtils,
-    routing::command::Command,
+    routing_api::command::Command,
     section::{SectionAuthorityProviderUtils, SectionUtils},
     Error, Result,
 };
