+++ src/routing/routing_api/enduser_registry.rs
