+++ src/routing/routing_api/tests/mod.rs
@@ -29,7 +29,7 @@ use crate::routing::{
     node::Node,
     peer::PeerUtils,
     relocation::{self, RelocatePayloadUtils, SignedRelocateDetailsUtils},
-    routing::core::{RESOURCE_PROOF_DATA_SIZE, RESOURCE_PROOF_DIFFICULTY},
+    routing_api::core::{RESOURCE_PROOF_DATA_SIZE, RESOURCE_PROOF_DIFFICULTY},
     section::{
         test_utils::*, ElderCandidatesUtils, NodeStateUtils, SectionAuthorityProviderUtils,
         SectionKeyShare, SectionPeersUtils, SectionUtils, FIRST_SECTION_MIN_AGE, MIN_ADULT_AGE,
