+++ src/routing/section/member_info.rs
@@ -7,7 +7,7 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use crate::messaging::node::{MemberInfo, Peer, PeerState};
-use crate::routing::routing::{error::Error, peer::PeerUtils};
+use crate::routing::routing_api::{error::Error, peer::PeerUtils};
 use xor_name::XorName;
 
 /// The minimum age a node can have. The Infants will start at age 4. This is to prevent frequent
