+++ src/routing/tests/utils/mod.rs
@@ -14,7 +14,7 @@ use anyhow::{bail, format_err, Error, Result};
 use ed25519_dalek::Keypair;
 use futures::future;
 use itertools::Itertools;
-use crate::routing::routing::{
+use crate::routing::routing_api::{
     Config, Event, EventStream, NodeElderChange, Routing, SectionAuthorityProviderUtils,
     TransportConfig, MIN_AGE,
 };
